Warnings in defines removed

+++ src/base/AS6802/CoRE4INET_AS6802Defs.h
@@ -23,7 +23,7 @@ namespace CoRE4INET {
 /**
  * Defines for range checks
  */
-#define MAX_CT_ID (uint16_t)0xFFFF
+#define MAX_CT_ID 0xFFFF
 #define MAX_PRIORITY 10
 #define MAX_BAG UINT32_MAX
 #define MAX_JITTER UINT32_MAX
